front/syn_metadata: move syn-walk metadata collectors out of front/ast (Slice 2)
Relocates collect_struct_names, collect_trait_names, collect_struct_origins,
and collect_immutable_field_attrs from front::ast into front::syn_metadata.
Callsites in front::ast::collect_program_metadata_pub and
lib::collect_struct_origin_metadata now reference the new module.

No behavioural change; pure module reshuffle of pure syn-walk helpers.
