+++ majit/majit-translate/src/front/ast.rs
@@ -123,8 +123,16 @@ pub fn collect_program_metadata_pub(parsed_files: &[ParsedInterpreter]) -> Progr
     let mut module_statics: HashMap<(String, String), crate::parse::ModuleStaticDecl> =
         HashMap::new();
     for parsed in parsed_files {
-        collect_struct_names(&parsed.file.items, "", &mut known_struct_names);
-        collect_trait_names(&parsed.file.items, "", &mut known_trait_names);
+        crate::front::syn_metadata::collect_struct_names(
+            &parsed.file.items,
+            "",
+            &mut known_struct_names,
+        );
+        crate::front::syn_metadata::collect_trait_names(
+            &parsed.file.items,
+            "",
+            &mut known_trait_names,
+        );
         // PyPy `annrpython.py` bookkeeper: every newly-seen STRUCT
         // gets cached under its lltype-object identity, which is the
         // defining-module path.  Pyre carries names as strings — record
@@ -133,7 +141,11 @@ pub fn collect_program_metadata_pub(parsed_files: &[ParsedInterpreter]) -> Progr
         // Empty `module_path` (legacy `parse_source` entry) skips the
         // record; consumers fall back to dual-publish convergence.
         if !parsed.module_path.is_empty() {
-            collect_struct_origins(&parsed.file.items, &parsed.module_path, &mut struct_origins);
+            crate::front::syn_metadata::collect_struct_origins(
+                &parsed.file.items,
+                &parsed.module_path,
+                &mut struct_origins,
+            );
         }
         // Mirror the per-file `ParsedInterpreter.use_imports` into the
         // program-wide `(module_path, alias) → fully_qualified_path`
@@ -179,146 +191,6 @@ pub fn collect_program_metadata_pub(parsed_files: &[ParsedInterpreter]) -> Progr
 /// Nested `mod foo { struct Bar; }` extends the prefix to `outer::foo`
 /// so the registered origin matches what `path_hash(canonical)` would
 /// produce for the qualified key.
-pub(crate) fn collect_struct_origins(
-    items: &[Item],
-    module_prefix: &str,
-    origins: &mut HashMap<String, String>,
-) {
-    for item in items {
-        match item {
-            Item::Struct(s) => {
-                let bare = s.ident.to_string();
-                // First-write-wins: if two files defined the same bare
-                // name, keep the first-seen.  Callers can disambiguate
-                // via use-import alias; the program-wide map only
-                // serves the most-common-case "bare name resolves to
-                // single module" path.
-                origins
-                    .entry(bare)
-                    .or_insert_with(|| module_prefix.to_string());
-            }
-            Item::Mod(m) => {
-                if let Some((_, ref sub_items)) = m.content {
-                    let nested = if module_prefix.is_empty() {
-                        m.ident.to_string()
-                    } else {
-                        format!("{}::{}", module_prefix, m.ident)
-                    };
-                    collect_struct_origins(sub_items, &nested, origins);
-                }
-            }
-            _ => {}
-        }
-    }
-}
-
-
-/// Read `#[jit_immutable_fields("a", "b?", "c[*]", "d?[*]")]` attributes
-/// off a struct declaration and return the declared field names paired
-/// with their `ImmutableRank`.  Bare idents (`#[jit_immutable_fields(a, b)]`)
-/// remain accepted as `ImmutableRank::Immutable` for backward compatibility.
-///
-/// Multiple attributes accumulate; non-recognised tokens are silently
-/// skipped (matching `syn::Meta::parse` looseness).  Rank suffix encoding
-/// follows RPython `rpython/rtyper/rclass.py:644-678 _parse_field_list`.
-fn collect_immutable_field_attrs(attrs: &[syn::Attribute]) -> Vec<(String, ImmutableRank)> {
-    use syn::punctuated::Punctuated;
-    use syn::{Expr, ExprLit, ExprPath, Lit, Token};
-
-    let mut specs = Vec::new();
-    for attr in attrs {
-        let Some(ident) = attr.path().get_ident() else {
-            continue;
-        };
-        if ident != "jit_immutable_fields" {
-            continue;
-        }
-        // Accept a comma-separated list of string literals and/or bare
-        // idents:  `#[jit_immutable_fields("foo?", bar)]`.  String form
-        // carries the RPython suffix; bare ident form is
-        // `ImmutableRank::Immutable`.
-        let parsed = attr.parse_args_with(Punctuated::<Expr, Token![,]>::parse_terminated);
-        let Ok(items) = parsed else {
-            continue;
-        };
-        for item in items {
-            match item {
-                Expr::Lit(ExprLit {
-                    lit: Lit::Str(s), ..
-                }) => {
-                    specs.push(ImmutableRank::parse(&s.value()));
-                }
-                Expr::Path(ExprPath { path, .. }) => {
-                    if let Some(id) = path.get_ident() {
-                        specs.push((id.to_string(), ImmutableRank::Immutable));
-                    }
-                }
-                _ => {}
-            }
-        }
-    }
-    specs
-}
-
-/// Pass 1a: collect all struct names (bare + qualified) recursively.
-fn collect_struct_names(
-    items: &[Item],
-    prefix: &str,
-    known_struct_names: &mut std::collections::HashSet<String>,
-) {
-    for item in items {
-        match item {
-            Item::Struct(s) => {
-                let bare_name = s.ident.to_string();
-                known_struct_names.insert(bare_name.clone());
-                if !prefix.is_empty() {
-                    known_struct_names.insert(format!("{}::{}", prefix, bare_name));
-                }
-            }
-            Item::Mod(m) => {
-                if let Some((_, ref sub_items)) = m.content {
-                    let mod_prefix = if prefix.is_empty() {
-                        m.ident.to_string()
-                    } else {
-                        format!("{}::{}", prefix, m.ident)
-                    };
-                    collect_struct_names(sub_items, &mod_prefix, known_struct_names);
-                }
-            }
-            _ => {}
-        }
-    }
-}
-
-fn collect_trait_names(
-    items: &[Item],
-    prefix: &str,
-    known_trait_names: &mut std::collections::HashSet<String>,
-) {
-    for item in items {
-        match item {
-            Item::Trait(trait_def) => {
-                let bare_name = trait_def.ident.to_string();
-                known_trait_names.insert(bare_name.clone());
-                if !prefix.is_empty() {
-                    known_trait_names.insert(format!("{}::{}", prefix, bare_name));
-                }
-            }
-            Item::Mod(m) => {
-                if let Some((_, ref sub_items)) = m.content {
-                    let mod_prefix = if prefix.is_empty() {
-                        m.ident.to_string()
-                    } else {
-                        format!("{}::{}", prefix, m.ident)
-                    };
-                    collect_trait_names(sub_items, &mod_prefix, known_trait_names);
-                }
-            }
-            _ => {}
-        }
-    }
-}
-
 /// Pass 1b: collect field types + fn return types using known_struct_names.
 ///
 /// `use_imports` is the per-source `use <path> as alias` table — same
@@ -366,7 +238,7 @@ fn collect_fields_and_returns(
                 // struct declaration (proc-macro pass-through in
                 // `majit_macros::jit_immutable_fields`). Multiple
                 // attributes accumulate.
-                let immutables = collect_immutable_field_attrs(&s.attrs);
+                let immutables = crate::front::syn_metadata::collect_immutable_field_attrs(&s.attrs);
                 if !immutables.is_empty() {
                     if prefix.is_empty() {
                         immutable_fields
